Added some logging from last night as I was desparately trying to figure out why this wasn't working.  Finally managed to get it to use single channel signed shorts for output, to match what gets generated for the .au file.  the problem was that I was still passing a float * to the fillBuffer: method.  Argh!  There still seem to be some scaling differences between the .au generation and TRMSynthesizer, and I don't understand why.
This has lots of extra code that needs to be cleaned up.

+++ Tools/softwareTRM/TRMSynthesizer.h
@@ -38,6 +38,6 @@ extern int verbose;
 - (void)stopPlaying;
 
 - (void)setupSoundDevice;
-- (BOOL)fillBuffer:(float *)buffer count:(int)count;
+- (BOOL)fillBuffer:(short *)buffer count:(int)count;
 
 @end
